+++ Maytaskfp3.ServiceBusTrigger/DTO.cs
@@ -1,0 +1,20 @@
+using System;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Host;
+using Microsoft.Extensions.Logging;
+using System.Text.Json;
+using System.Net.Http;
+using System.Threading.Tasks;
+
+
+
+namespace Maytaskfp3.ServiceBusTrigger
+{
+    public class Qmodel
+    {
+        public string message { get; set; }
+    }
+ }
+    
+
+
