+++ Maytaskfp3.ServiceBusTrigger/Startup.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.Azure.Functions.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http.Headers;
+using System.Net.Http;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+
+[assembly: FunctionsStartup(typeof(Maytaskfp3.ServiceBusTrigger.Startup))]
+
+namespace Maytaskfp3.ServiceBusTrigger
+{
+    public class Startup : FunctionsStartup
+    {
+        public override void Configure(IFunctionsHostBuilder builder)
+        {
+            builder.Services.AddHttpClient("Httpclient");
+        }
+    }
+}
